init uebung3

+++ Algorithmen/uebung3.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void rectangleMethod(float _a, float _b, int _i);
-void trapezoidMethod(float _a, float _b, int _i);
-void simpsonMethod(float _a, float _b, int _i);
+void rectangleMethod(float _a, float _b, float _i);
+void trapezoidMethod(float _a, float _b, float _i);
+void simpsonMethod(float _a, float _b, float _i);
